Move Admin API to `admin` sub module, use enums, fix doc strings
