+++ setup.py
@@ -27,7 +27,8 @@ setup(name='confluent-kafka',
       packages=find_packages(exclude=("tests", "tests.*")),
       data_files=[('', ['LICENSE.txt'])],
       install_requires=[
-          'futures;python_version<"3.0"'
+          'futures;python_version<"3.0"',
+          'enum34;python_version<"3.0"'
       ],
       extras_require={
           'avro': ['fastavro', 'requests', avro]
